Moved all http source/headers to their own folders
