+++ tests/CMakeLists.txt
@@ -10,7 +10,7 @@ set(TEST_HEADER_LIST
 
 set(TEST_SOURCE_LIST
         main.c
-        ../include/http_constants.h)
+        ../include/http/http_constants.h)
 
 include_directories(${CGREEN_PUBLIC_INCLUDE_DIRS} ${PROJECT_BINARY_DIR})
 add_executable(template_test
